Refactor api DI of service
